+++ dag!/src/main/java/com/qicubo/mobile/dag/services/BolhaService.java
@@ -1,5 +1,6 @@
 package com.qicubo.mobile.dag.services;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.qicubo.mobile.dag.models.Bolha;
@@ -9,7 +10,7 @@ public interface BolhaService extends Service<Bolha>{
 	     
     List<Bolha> findBolhaByUser(Usuario user);
     
-    List<Bolha> findAllCloserBolhas(Long latitude, Long longitude);
+    List<Bolha> findAllCloserBolhas(BigDecimal latitude, BigDecimal longitude);
     
     public boolean isBolhaExist(Bolha bolha);
